- Aggiornamento composer a PHP 7.4; - Aggiunto pulsante (non fa nulla) per download computo metrico; - Aggiunto pulsante (apre modale) per creazione del computo metrico;

+++ rossiduenord/resources/views/livewire/users/index.blade.php
@@ -56,33 +56,35 @@
 								        name="pencil-alt"
 								        class="w-5 h-5 cursor-pointer text-indigo-500 hover:text-indigo-800"></x-icon>
 
-								<x-modal>
-									<x-slot name="trigger">
-										<div class="text-red-600 hover:text-red-900">
-											<x-icon name="trash" class="w-5 h-5"></x-icon>
-										</div>
-									</x-slot>
-									<x-slot name="title">
-										Conferma eliminazione
-									</x-slot>
-									Sei sicuro di voler eliminare l'utente <span
-											class="font-bold">{{ $user->name }}</span>?
-									<x-slot name="footer">
-										<x-link-button x-on:click="open = false">Annulla</x-link-button>
-										<x-danger-button class="ml-2" wire:click="deleteUser({{ $user->id }})"
-										                 wire:loading.attr="disabled">
-											Elimina
-										</x-danger-button>
-									</x-slot>
-								</x-modal>
-								@can('impersonate_users')
-									@canImpersonate($guard = null)
-									<a href="{{ route('impersonate', $user->id) }}">
-										<x-icon name="sparkles"
-										        class="w-5 h-5 cursor-pointer text-yellow-500 hover:text-yellow-800"></x-icon>
-									</a>
-									@endCanImpersonate
-								@endcan
+								@if($user->id !== auth()->user()->id)
+									<x-modal>
+										<x-slot name="trigger">
+											<div class="text-red-600 hover:text-red-900">
+												<x-icon name="trash" class="w-5 h-5"></x-icon>
+											</div>
+										</x-slot>
+										<x-slot name="title">
+											Conferma eliminazione
+										</x-slot>
+										Sei sicuro di voler eliminare l'utente <span
+												class="font-bold">{{ $user->name }}</span>?
+										<x-slot name="footer">
+											<x-link-button x-on:click="open = false">Annulla</x-link-button>
+											<x-danger-button class="ml-2" wire:click="deleteUser({{ $user->id }})"
+											                 wire:loading.attr="disabled">
+												Elimina
+											</x-danger-button>
+										</x-slot>
+									</x-modal>
+									@can('impersonate_users')
+										@canImpersonate($guard = null)
+										<a href="{{ route('impersonate', $user->id) }}">
+											<x-icon name="sparkles"
+											        class="w-5 h-5 cursor-pointer text-yellow-500 hover:text-yellow-800"></x-icon>
+										</a>
+										@endCanImpersonate
+									@endcan
+								@endif
 							</div>
 						@endif
 					</x-table.td>
